requirements.txt
streamlit==1.31.0
groq==0.4.2
faiss-cpu==1.9.0.post1
sentence-transformers==2.3.1
numpy<2.0
httpx==0.27.2
pandas
pyarrow
requests